test_ajax5.html improves insertWebworkIframe to take arguments  (option, path, width, height) where options is a json block with all of the query segments needed to call the problem.
Still to do: convert all of the inputs to named inputs so it is easier to leave one of them out.

Next: write function that will request a username and password and store it in sessionStorage and modify insertWebworkIframe to use
those values for userID and course_password.    (Can we just use the session_key instead of password? need to check server side behavior
it was supposed to accept either a password or a session key.)